Unlink people translations from roles
The current logic in Whitehall is to only show the available translations where: 1) translations for the person are present AND 2) translations for every role they have are present.

I'm not sure why the two need to be linked, and it's confusing when translations are present in Whitehall but don't appear on the live site.