Consistent Capybara DSL usage
This updates Capybara based assertions to favour using Capybaras built
in assertion methods (for example assert_selector, assert_text,
assert_no_text) and makes an opinionated call to avoid using the `page.`
approach unless necessary.

The use of Capybara assertions means that more information is provided
when these fail compared to the standard `assert` method which just
tests against a boolean.

The switch to avoid `page.` is for consistency as there are lots of
fluctuations in the usage of Capybara DSL.