Handle Brakeman warning: Parameters should be whitelisted for mass assignment
Brakeman now doesn't allow implicitly permitting all params with a bang
and wants us to be explicit with every parameter we're permitting. This
makes sense but unfortunately whitehall doesn't really follow this
protocol in quite a few places so this and the following commits
implements explicitly requiring params.

This requires params for the admin interfaces within whitehall, I've
looked through the code and am pretty confident all the potential params
are covered.