Fix the Publishing API HTML Attachment handling
This comes from investigating issues where an edition was unpublished,
but the HTML attachments were not.

It's possible for this to break because the worker is enqueued during
the transaction that creates the entry in the unpublishings table. If
the worker runs before this transaction commits, then it won't
error (due to an odd guard condition), it'll just fail silently.

This commit changes the PublishingApiHtmlAttachmentsWorker in to a
normal class. This means that the code runs within the transaction, so
there shouldn't be any issues accessing the unpublishing.

There are some editions with high numbers (up to 99) of HTML
attachments, so the actual calls to the Publishing API to handle these
are still offloaded through Sidekiq jobs.