Support creating drafts from some superseded editions
There are a number of documents in Whitehall where the latest edition
is in the superseded state, a bit of a contradiction. For the user,
this means there's nothing they can do any more with this document.

I believe a large proportion of these cases seem to be because the
published edition was made a draft when unpublished, and then somehow
deleted.

To try and help with this situation, this commit makes it possible to
create a new draft for documents in this state. The button appears on
the page for the latest edition, which is superseded. The draft
however is based on the oldest deleted edition which is newer than the
superseded edition, as this will probably have been published in the
past, and is most likely where the updated content is.

Of course, this is very uncertain, part of this is due to not being
able to really identify why these documents are in a suposedly
impossible state. It does however offer a better user experience.