Test create draft from first deleted draft following superseded edition
There are a number of documents in Whitehall for which the
latest edition is also superseded, something of a
contradiction.

To allow for a way out in this circumstance, find the deleted
edition that was probably the one that supersedes this
superseded edition, and use that to create the draft.