Fix marshalling errors
We were seeing errors in Logit as follows:

```
[4938944e-2f85-4a07-902f-585841876d22] Marshalling error for key 'count=1000/fields=["display_type", "title", "link", "public_timestamp", "format", "content_store_document_type", "description", "content_id", "organisations", "document_collections"]/filter_topical_events=coronavi:md5:a392a8c87eaf0fcfcbf88dcf77d910aa': no _dump_data is defined for class Thread::Mutex
```

The fields correspond to the `Rails.cache.fetch` call in
SearchRummagerService. The error appears to be because we're
attempting to serialize a complex object [1].

We need to implement our own `marshal_dump` and `marshal_load`
methods to describe to the Marshal how to serialise and unserialise
our object. [2]

Luckily the `RummagerDocumentPresenter` object returned in our
`SearchRummagerService` already has a `as_hash` method, and takes
a hash in its `initialize` method, so the serialisation work has
already been done.

Trello: https://trello.com/c/ZZCFTp52/1974-5-investigate-whitehall-marshalling-errors

[1]: https://stackoverflow.com/a/5753365
[2]: https://ruby-doc.org/core-2.7.1/Marshal.html#module-Marshal-label-marshal_dump+and+marshal_load