Add base tag to html attachments view
This commit fixes a long-standing bug with Whitehall, where links in the header and footer of the CSV preview page (rendered by the `CsvPreviewController`) contain relative links that point to the incorrect site root (`assets.publishing.service.gov.uk` rather than `www.gov.uk`), resulting in `404`s to any of those links which are followed. The footer itself is obtained through Slimmer, using the `chromeless` view that's part of `Static` where most URLs in the footer are set as relative.

The fix is achieved through adding a `base` tag to the `frontend_base` partial, which is set from the `CsvPreviewController` and rendered via the `html_attachments` view.

By adding a `base` tag, all relative URLs on the page will be prefixed with the `href` attribute specified, which is the GOV.UK website root. The `base` tag can only appear in the `head` of a page, which is the reason for adding it into the partial `frontend_base`.