Create a default rake task that lints and runs tests
This offers a close approximation as to what runs in CI for Whitehall.
Most GOV.UK apps use the default rake task to run CI checks however
Whitehall is special in this regard, this task is added to provide a
higher degree of consistency.

These tests seem to require assets:precompile to be run before the test
suite otherwise one test fails due to [1].

[1]: https://github.com/alphagov/whitehall/commit/ec98c5045eee19e17ac1e7edb82dc86cb8eeb15c