Ignore !important declarations
This ignores the !important declarations on an individual basis. I
decided not to apply this as a global rule as mostly these were applied
only to special cases rather than being a normal rule and I hope that
having this as an offence makes any future developers reconsider using
important rather than consider adding another ignore.