Make naming of SCSS files consistent
This apply a consistent naming approach across all of the SCSS files in
this project. It consistently applies an underscore prefix for every
partial and applies a consistent naming approach (kebab-case) to every
file.

This naming convention is a little awkward given that we have sections
of SCSS that are named based on views and thus require switching Ruby
underscore convention to a dash approach. It struck me as better doing
this than having an arbitrary switch when dealing with view directories.